fixed no build/ dir setup error

+++ setup.py
@@ -8,11 +8,12 @@ __location__ = os.path.realpath(
 
 # Remove existing builds already in build/lib
 build_path = os.path.join(__location__, "build")
-for folder in os.listdir(build_path):
-    if folder.startswith("lib."):
-        folder_path = os.path.join(build_path, folder)
-        for archive in os.listdir(folder_path):
-              os.remove(os.path.join(folder_path, archive))
+if os.path.isfile(build_path):
+    for folder in os.listdir(build_path):
+        if folder.startswith("lib."):
+            folder_path = os.path.join(build_path, folder)
+            for archive in os.listdir(folder_path):
+                os.remove(os.path.join(folder_path, archive))
 
 # Compile extension
 module = Extension(module_name, sources=["src/quantum.c"])
